Conflicts:
	.gitignore
	arena/thunderdome/templates/thunderdome/base.html

+++ masterblaster/analyzer.py
@@ -10,6 +10,7 @@ import math
 import gzip
 import os
 import urllib2
+import argparse
 from copy import copy
 from datetime import datetime, timedelta
 from collections import defaultdict
@@ -26,12 +27,16 @@ import django
 django.setup()
 
 def main():
-    #############TEMPORARY REMOVE
-    for x in Game.objects.all():
-        x.score = -1
-        x.save()
-    for x in GameStats.objects.all():
-        x.delete()
+    parser = argparse.ArgumentParser(description='Game Analyzer')
+    parser.add_argument('--r', action='store_true', help='Reset all scores and start over')
+    args = parser.parse_args()
+    print args
+    if args.r:
+        for x in Game.objects.all():
+            x.score = -1
+            x.save()
+        for x in GameStats.objects.all():
+            x.delete()
     try:
         gamestats = GameStats.objects.get(game=game_name)
     except ObjectDoesNotExist:
@@ -55,12 +60,13 @@ def main():
     gamestats.save()
     while gamestats.numPlayed < 100:
         for x in Game.objects.all().filter(score=-1).filter(status="Complete"):
-            if x.gamelog_url == '':
+            url = x.gamelog_url
+            try:
+                u = urllib2.urlopen(url)
+            except:
                 x.score = -2
                 x.save()
                 continue
-            url = x.gamelog_url
-            u = urllib2.urlopen(url)
             meta = u.info()
             file_size = int(meta.getheaders("Content-Length")[0])
             print "Game", x, "is", file_size, "bytes"
@@ -89,12 +95,14 @@ def main():
 
 def analyse_game(game):
     gamestats = GameStats.objects.get(game=game_name)
-    if game.gamelog_url == '':
+    url = game.gamelog_url
+    try:
+        u = urllib2.urlopen(url)
+    except:
         game.score = -2
         game.save()
+        print "Invalid gamelog url"
         return
-    url = game.gamelog_url
-    u = urllib2.urlopen(url)
     meta = u.info()
     file_size = int(meta.getheaders("Content-Length")[0])
     if file_size > gamestats.maxSize:
@@ -103,7 +111,7 @@ def analyse_game(game):
     gamestats.save()
     
     #Begin scoring
-    if file_size > (gamestats.maxSize * 0.3) and file_size < (gamestats.maxSize * 0.8):
+    if file_size > (gamestats.maxSize * 0.2) and file_size < (gamestats.maxSize * 0.7):
         game.score += 1
         print "Game", game, "is an interesting length!"
     if not game.discon_happened:
